Updates to Assign3: - fixed error with ^ command not verifying argument is a number - added comments to ^ command function

+++ prev. assns/Assign4/src/PiComputer.java
@@ -1,16 +1,18 @@
 /**
- * Class description for a Pi Computer used for spawning as a thread to retrieve and 
+ * Class description for a Pi Computer used for spawning as a thread to retrieve and
  * execute tasks from the task queue.
  *
  * @author Ky Kartchner
  */
 public class PiComputer implements Runnable {
-    /** For keeping track of how many tasks have been executed among all the threads */
+    /**
+     * For keeping track of how many tasks have been executed among all the threads
+     */
     private static int completionCount = 0;
 
     private TaskQueue taskQueue;
     private ResultTable resultTable;
-    
+
 
     PiComputer(TaskQueue queue, ResultTable table) {
         this.taskQueue = queue;
@@ -19,18 +21,18 @@ public class PiComputer implements Runnable {
 
     /**
      * Pops a task from the task queue, executes it, then stores the result into the result table;
-     * Prints a period whenever the number of total tasks completed (completionCount) is a 
+     * Prints a period whenever the number of total tasks completed (completionCount) is a
      * multiple of 10, and a newline every time it is a multiple of 200.
      */
     @Override
     public void run() {
         final int PROGRESS_MULTIPLE = 10;
         final int BREAK_MULTIPLE = 200;
-       
+
         Task task;
         while ((task = taskQueue.pop()) != null) { // While the task that was popped is not null
             try {
-                resultTable.storeValue(task.getId(), (Integer) task.perform()); // Compute and store value
+                resultTable.storeValue(task.getId(), task.perform()); // Compute and store value
                 if (++completionCount % PROGRESS_MULTIPLE == 0) { // Print period every multiple of 10
                     System.out.print(".");
                     if (completionCount % BREAK_MULTIPLE == 0) { // New line every multiple of 200
